+++ CHANGELOG.md
@@ -9,3 +9,14 @@
 - Different physic shader support.
 - Physics engine with per-object collision types (Box/Sphere).
 - Same speed as pure Vulkano+winit
+
+## [0.1.1] - 2026-04-1
+
+### Added
+
+- Collision types as enum (Sphere, Box).
+- Optional apply one physic/visual shader for all object on scene
+
+### Fixed
+
+- Better performance( main loop refactoring )